feat: extract notification icon and color functions into separate utility module for better organization
