+++ lib/index.ts
@@ -10,3 +10,4 @@ export { deleteFileFromR2, getPresignedDownloadUrl, uploadFileToR2 } from "./r2-
 export { toTitleCase } from "./functions/text/title-case";
 export { baseFormConfig, createFormConfig } from "./config/hook-form-config";
 export { getCurrentUtcDate, getCurrentUtcISOString } from "./datetime";
+export { getNotificationColor, getNotificationIcon } from "./notification-ui-utils";
